training/requirements.txt
torch>=2.5.0
numpy>=1.26.0
h5py>=3.10.0
geopandas>=0.14.0
shapely>=2.0.0
pyvista>=0.43.0
laspy>=2.5.0
onnx>=1.15.0
onnxruntime>=1.17.0
pyyaml>=6.0.1
tensorboard>=2.15.0
rasterio>=1.3.0
scipy>=1.12.0
